+++ integrations/rollup/tests/integration.test.js
@@ -138,7 +138,7 @@ describe('watcher', () => {
   })
 
   test(`classes are generated when the index.css file changes`, async () => {
-    await writeInputFile('index.html', html`<div class="font-bold btn"></div>`)
+    await writeInputFile('index.html', html`<div class="btn font-bold"></div>`)
 
     let runningProcess = $('rollup -c --watch')
     await runningProcess.onStderr(ready)
@@ -160,7 +160,7 @@ describe('watcher', () => {
 
         @layer components {
           .btn {
-            @apply px-2 py-1 rounded;
+            @apply rounded px-2 py-1;
           }
         }
       `
@@ -191,7 +191,7 @@ describe('watcher', () => {
 
         @layer components {
           .btn {
-            @apply px-2 py-1 rounded bg-red-500;
+            @apply rounded bg-red-500 px-2 py-1;
           }
         }
       `