+++ integrations/vite/tests/integration.test.js
@@ -169,7 +169,7 @@ describe('watcher', () => {
       'index.html',
       html`
         <link rel="stylesheet" href="./index.css" />
-        <div class="font-bold btn"></div>
+        <div class="btn font-bold"></div>
       `
     )
 
@@ -193,7 +193,7 @@ describe('watcher', () => {
 
         @layer components {
           .btn {
-            @apply px-2 py-1 rounded;
+            @apply rounded px-2 py-1;
           }
         }
       `
@@ -224,7 +224,7 @@ describe('watcher', () => {
 
         @layer components {
           .btn {
-            @apply px-2 py-1 rounded bg-red-500;
+            @apply rounded bg-red-500 px-2 py-1;
           }
         }
       `